Strip algo, meta and util procedures from full
