+++ core/src/main/java/apoc/algo/Cover.java
