+++ core/src/main/java/apoc/algo/PathFinding.java
@@ -1,72 +1,22 @@
 package apoc.algo;
 
-import apoc.path.RelationshipTypeAndDirections;
 import apoc.result.PathResult;
 import apoc.result.WeightedPathResult;
 import apoc.util.Util;
 import org.neo4j.graphalgo.*;
 import org.neo4j.graphdb.*;
-import org.neo4j.internal.helpers.collection.Pair;
 import org.neo4j.procedure.Context;
 import org.neo4j.procedure.Description;
 import org.neo4j.procedure.Name;
 import org.neo4j.procedure.Procedure;
-import org.neo4j.values.storable.PointValue;
 
 import java.util.Collections;
 import java.util.Map;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
+import static apoc.algo.PathFindingUtils.buildPathExpander;
 
 public class PathFinding {
-
-    public static class GeoEstimateEvaluatorPointCustom implements EstimateEvaluator<Double> {
-
-        // -- from org.neo4j.graphalgo.impl.util.GeoEstimateEvaluator
-        private static final double EARTH_RADIUS = 6371 * 1000; // Meters
-        private Node cachedGoal;
-        private final String pointPropertyKey;
-        private double[] cachedGoalCoordinates;
-        
-        public GeoEstimateEvaluatorPointCustom(String pointPropertyKey) {
-            this.pointPropertyKey = pointPropertyKey;
-        }
-        
-        @Override
-        public Double getCost( Node node, Node goal) {
-            double[] nodeCoordinates = getCoordinates(node);
-            if ( cachedGoal == null || !cachedGoal.equals( goal ) )
-            {
-                cachedGoalCoordinates = getCoordinates(goal);
-                cachedGoal = goal;
-            }
-            return distance(nodeCoordinates[0], nodeCoordinates[1],
-                    cachedGoalCoordinates[0], cachedGoalCoordinates[1] );
-        }
-        
-        private static double distance( double latitude1, double longitude1,
-                                        double latitude2, double longitude2 ) {
-            latitude1 = Math.toRadians( latitude1 );
-            longitude1 = Math.toRadians( longitude1 );
-            latitude2 = Math.toRadians( latitude2 );
-            longitude2 = Math.toRadians( longitude2 );
-            double cLa1 = Math.cos( latitude1 );
-            double xA = EARTH_RADIUS * cLa1 * Math.cos( longitude1 );
-            double yA = EARTH_RADIUS * cLa1 * Math.sin( longitude1 );
-            double zA = EARTH_RADIUS * Math.sin( latitude1 );
-            double cLa2 = Math.cos( latitude2 );
-            double xB = EARTH_RADIUS * cLa2 * Math.cos( longitude2 );
-            double yB = EARTH_RADIUS * cLa2 * Math.sin( longitude2 );
-            double zB = EARTH_RADIUS * Math.sin( latitude2 );
-            return Math.sqrt( ( xA - xB ) * ( xA - xB ) + ( yA - yB )
-                    * ( yA - yB ) + ( zA - zB ) * ( zA - zB ) );
-        }
-        // -- end from org.neo4j.graphalgo.impl.util.GeoEstimateEvaluator
-
-        private double[] getCoordinates(Node node) {
-            return ((PointValue) node.getProperty(pointPropertyKey)).coordinate();
-        }
-    }
 
     @Context
     public GraphDatabaseService db;
@@ -108,7 +58,7 @@ public class PathFinding {
         String pointPropertyName = (String) config.get("pointPropName");
         final EstimateEvaluator<Double> estimateEvaluator;
         if (pointPropertyName != null) {
-            estimateEvaluator = new GeoEstimateEvaluatorPointCustom(pointPropertyName);
+            estimateEvaluator = new PathFindingUtils.GeoEstimateEvaluatorPointCustom(pointPropertyName);
         } else {
             String latPropertyName = config.getOrDefault("y", "latitude").toString();
             String lonPropertyName = config.getOrDefault("x", "longitude").toString();
@@ -158,27 +108,6 @@ public class PathFinding {
         Iterable<Path> allPaths = algo.findAllPaths(startNode, endNode);
         return StreamSupport.stream(allPaths.spliterator(), false)
                 .map(PathResult::new);
-    }
-
-    public static PathExpander<Double> buildPathExpander(String relationshipsAndDirections) {
-        PathExpanderBuilder builder = PathExpanderBuilder.empty();
-        for (Pair<RelationshipType, Direction> pair : RelationshipTypeAndDirections
-                .parse(relationshipsAndDirections)) {
-            if (pair.first() == null) {
-                if (pair.other() == null) {
-                    builder = PathExpanderBuilder.allTypesAndDirections();
-                } else {
-                    builder = PathExpanderBuilder.allTypes(pair.other());
-                }
-            } else {
-                if (pair.other() == null) {
-                    builder = builder.add(pair.first());
-                } else {
-                    builder = builder.add(pair.first(), pair.other());
-                }
-            }
-        }
-        return builder.build();
     }
 
 }
