+++ core/src/main/java/apoc/export/arrow/ExportResultStrategy.java
@@ -1,6 +1,6 @@
 package apoc.export.arrow;
 
-import apoc.meta.Meta;
+import apoc.meta.Types;
 import apoc.util.Util;
 import org.apache.arrow.vector.types.pojo.Field;
 import org.apache.arrow.vector.types.pojo.Schema;
@@ -32,7 +32,7 @@ public interface ExportResultStrategy {
     default Schema schemaFor(GraphDatabaseService db, List<Map<String, Object>> records) {
         final List<Field> fields = records.stream()
                 .flatMap(m -> m.entrySet().stream())
-                .map(e -> new AbstractMap.SimpleEntry<>(e.getKey(), fromMetaType(Meta.Types.of(e.getValue()))))
+                .map(e -> new AbstractMap.SimpleEntry<>(e.getKey(), fromMetaType(Types.of(e.getValue()))))
                 .collect(Collectors.groupingBy(e -> e.getKey(), Collectors.mapping(e -> e.getValue(), Collectors.toSet())))
                 .entrySet()
                 .stream()
