+++ core/src/main/java/apoc/path/PathExplorer.java
