+++ core/src/test/java/apoc/algo/PathFindingTest.java
