+++ core/src/test/java/apoc/cypher/CypherInitializerTest.java
