+++ core/src/test/java/apoc/path/LabelSequenceTest.java
