+++ core/src/test/java/apoc/path/NodeFilterTest.java
