+++ core/src/test/java/apoc/path/PathsTest.java
