+++ core/src/test/java/apoc/path/SequenceTest.java
