+++ full/src/test/java/apoc/ApocSplitTest.java
@@ -23,8 +23,6 @@ public class ApocSplitTest {
         "apoc.trigger.list",
         "apoc.trigger.pause",
         "apoc.trigger.resume",
-        "apoc.util.sleep",
-        "apoc.util.validate",
         "apoc.merge.node.eager",
         "apoc.merge.node",
         "apoc.merge.nodeWithStats.eager",
@@ -41,11 +39,6 @@ public class ApocSplitTest {
         "apoc.nodes.collapse",
         "apoc.nodes.group",
         "apoc.example.movies",
-        "apoc.path.expand",
-        "apoc.path.expandConfig",
-        "apoc.path.subgraphNodes",
-        "apoc.path.subgraphAll",
-        "apoc.path.spanningTree",
         "apoc.graph.fromData",
         "apoc.graph.from",
         "apoc.graph.fromPath",
@@ -59,21 +52,6 @@ public class ApocSplitTest {
         "apoc.lock.read.nodes",
         "apoc.lock.rels",
         "apoc.lock.read.rels",
-        "apoc.algo.aStar",
-        "apoc.algo.aStarConfig",
-        "apoc.algo.dijkstra",
-        "apoc.algo.allSimplePaths",
-        "apoc.algo.cover",
-        "apoc.meta.stats",
-        "apoc.meta.data.of",
-        "apoc.meta.data",
-        "apoc.meta.schema",
-        "apoc.meta.nodeTypeProperties",
-        "apoc.meta.relTypeProperties",
-        "apoc.meta.graph",
-        "apoc.meta.graph.of",
-        "apoc.meta.graphSample",
-        "apoc.meta.subGraph",
         "apoc.cypher.runTimeboxed",
         "apoc.cypher.run",
         "apoc.cypher.runMany",
@@ -146,14 +124,6 @@ public class ApocSplitTest {
         "apoc.temporal.format",
         "apoc.temporal.formatDuration",
         "apoc.temporal.toZonedTemporal",
-        "apoc.util.sha1",
-        "apoc.util.sha256",
-        "apoc.util.sha384",
-        "apoc.util.sha512",
-        "apoc.util.md5",
-        "apoc.util.validatePredicate",
-        "apoc.util.decompress",
-        "apoc.util.compress",
         "apoc.node.relationship.exists",
         "apoc.nodes.connected",
         "apoc.node.labels",
@@ -190,10 +160,6 @@ public class ApocSplitTest {
         "apoc.date.convertFormat",
         "apoc.date.add",
         "apoc.label.exists",
-        "apoc.meta.cypher.isType",
-        "apoc.meta.cypher.type",
-        "apoc.meta.cypher.types",
-        "apoc.meta.nodes.count",
         "apoc.diff.nodes",
         "apoc.cypher.runFirstColumnMany",
         "apoc.cypher.runFirstColumnSingle",
